requirements.txt
aiofiles==0.8.0
fastapi==0.75.0
jinja2==3.0.3
numpy==1.19.5
pandas==1.1.5
uvicorn==0.16.0
psutil==5.9.0
cryptography==37.0.2
python-dateutil
python-dotenv